msm: ipa: ipa3_ctx is defined in ipa3_tx_dp()
This patch simply deletes a check whether ipa3_ctx is null when
ipa3_tx_dp() is called.  It's a trivial change, but the explanation
below proves that it's correct.

Only ipa3_wwan_xmit() calls ipa3_tx_dp().  That function is the
ndo_start_xmit field in the rmnet IPA network device operations
structure.  This operation won't be used until it has been supplied
to alloc_netdev() via the net_device structure, in ipa3_wwan_probe().
And that call to alloc_netdev() happens only after ipa3_is_ready()
returns true, which can only happen if ipa3_ctx is non-null.

So ipa3_ctx is known to be non-null by the time ipa3_wwan_xmit()
is registered, so it will be non-null when that function is called.

Change-Id: I4757ea8df89cef3f6d04d54c1c982e771775f8f8
Signed-off-by: Alex Elder <elder@codeaurora.org>