msm: ipa: ipa3_ctx is non-null in ipa3_post_init()
Delete a check whether ipa3_ctx is null when ipa3_post_init() is
called.  The explanation below explains why it's correct.

The only thing that calls ipa3_post_init() is ipa3_post_init_wq().
That function is only called when the ipa3_post_init_work work item
is queued, which occurs only in ipa3_write(), which is the ->write
file operation for the IPA character device.

Nothing can call the file write operation for a device before
cdev_init() is called to initialize the the IPA character device
in ipa3_pre_init().  And by the time that is called, ipa3_ctx
already points to non-null context structure.

Therefore ipa3_post_init() does not need to verify ipa3_ctx is
non-null, and that check can be removed.

Change-Id: Ibf786ceee1cb8c01ca9d795531e9d88622d5b4cc
Signed-off-by: Alex Elder <elder@codeaurora.org>