msm: ipa: directly supply device name for debug
All of the IPAERR*(), IPAWANERR*(), and IPAHAL_ERR*() macros insert a
device/driver name by string concatenation before the start of the
format string supplied to (e.g.) pr_err() or IPA_IPC_LOGGING().
This doesn't really add value; the IPA, IPAWAN, and IPAHAL debug
functions should always show "ipa", "ipa-wan", and "iphal",
respectively.

Just insert the string values explicitly at the beginning of the
format string.  This is setting up so we can use pr_fmt() in calls
to pr_err() and related functions.  In the case of the HAL related
functions, this eliminates the need for IPAHAL_DRV_NAME.

There is one outlier in "rmnet_ipa.c", which follows the same
pattern as is used in the pr_err() calls in the IPADBG() macro.
Fix that instance together with the reset.

Change-Id: I90c5a6b08c0f4d8c9a89919079bc5d1276bf290d
Signed-off-by: Alex Elder <elder@codeaurora.org>