msm: ipahal: use pr_fmt() for HAL pr_*() calls
The normal way to consistently modify format strings used for
pr_debug(), pr_err(), etc. is to define a macro pr_fmt() at the
top of source files that use them.

In the IPA HAL code, the only place these functions are used is in
the definitions of IPAHAL_DBG(), IPAHAL_DBG_LOW(), and IPAHAL_ERR().

Define pr_fmt() in a way that will produce the same result as
occurs currently in the source files that call the IPAHAL_DBG()
(etc.) macros.  Change the pr_err() and pr_debug() calls used in
the macros so they assume the use of pr_fmt().

Change-Id: Id4e7e26d6e937d95be4c79a644500557e2f62e81
Signed-off-by: Alex Elder <elder@codeaurora.org>