msm: ipa-wan: use pr_fmt() for pr_*() calls
The normal way to consistently modify format strings used for
pr_debug(), pr_err(), etc. is to define a macro pr_fmt() at the
top of source files that use them.

In the IPA WAN code, just about the only place these functions
are used is in the definitions of IPAWANDBG(), IPAWANDBG_LOW(),
IPAWANERR(), and IPAWANINFO().  (The one other spot is a call
to pr_err_ratelimited() in apps_ipa_packet_receive_notify(), which
we fix here as well.)

Define pr_fmt() in a way that will produce the same result as occurs
currently in the source files that call the IPAHAL_DBG() (etc.)
macros.  Change the pr_err() and pr_debug() calls used in the macros
so they assume the use of pr_fmt().

Change-Id: I41ed4f123fff67d4faa4618b1c7a26d293572670
Signed-off-by: Alex Elder <elder@codeaurora.org>