msm: ipa: use pr_fmt() for pr_*() calls
The normal way to consistently modify format strings used for
pr_debug(), pr_err(), etc. is to define a macro pr_fmt() at the top
of source files that use them.

In the IPA code, these functions are used is in the definitions of
IPADBG(), IPADBG_LOW(), IPAERR(), and IPAERR_RL().

Define pr_fmt() in a way that will produce the same result as occurs
currently in the source files that call the IPAHAL_DBG() (etc.)
macros.  Change the pr_err() and pr_debug() calls used in the macros
so they assume the use of pr_fmt().

Change-Id: Ifa74d01e9113a18a473dbea18d487ae90b73498a
Signed-off-by: Alex Elder <elder@codeaurora.org>