msm: ipa-wan: open-code pr_*() function calls
The next patch is going to switch to using pr_fmt() for manipulating
the format string on calls to pr_err(), pr_debug(), etc. in the IPA
WAN code.  In order to avoid that changing the behavior of the
current code, change all current calls to pr_err() (etc.) to direct
calls to printk().  This is just open-coding the pr_*() macros while
preserving functionality.

Change-Id: Iccd0ece8028befc986f2016ae7e244192d270fc6
Signed-off-by: Alex Elder <elder@codeaurora.org>