msm: ipa: fix ipa3_is_ready()
The function ipa3_is_ready() indicates whether IPA is "ready".
Whether it is ready is recorded in a field in the IPA context,
ipa_initialization_complete.  Because ipa3_is_ready() can be called
before the IPA context has been initialized, it must also check the
validity of the ipa3_ctx pointer before checking the state.

The function that makes the IPA ready (and sets the "ready" flag)
is ipa3_post_init().  That function is called only via workqueue,
which is only activated by ipa3_write() after firmware has been
loaded.  The top of ipa3_post_init() calls ipa3_is_ready(), but
there is a race between checking that flag at the top of the
function and setting it at the end, so despite that check it is
possible for multiple processes to be executing that function
concurrently.

Furthermore, the check for "ready" should occur earlier than when
ipa3_post_init() is called.  Instead, ipa3_write() should be doing
this, to avoid concurrent attempts to get firmware.  (It seems the
firmware loading code may be managing such potential concurrency,
but it's better to avoid that to begin with.)

This patch fixes two problems.  First, it moves the check to avoid
concurrency to the top if ipa3_write(), and removes it from
ipa3_post_init().  That avoids concurrent attempts to load firmware,
which also ensures the ipa3_post_init() gets initiated only once
(but that depends on fix for the second problem).

The second problem fixed is that we now have three IPA states:
INITIAL; STARTING; and READY.  The top of ipa3_write() checks (and
possibly sets) the state atomically.  Only if the state is INITIAL
will the current process load firmware; in that case the state
(atomically) becomes STARTING, blocking out any other processes.
If successful, ipa3_write() arranges for ipa3_post_init() to be
called, and when succeeds the IPA state will become READY.  Failures
while in STARTING state cause the state to return to INITIAL.

The old ipa_initialization_complete field is removed, replaced with
an atomic state field.

Change-Id: Ibc29b9a2d9017fd0bba4ddefe45e48624ef157e3
Signed-off-by: Alex Elder <elder@codeaurora.org>