msm: ipa: pass flag to ipa_ipc_logging()
Only two log buffers are ever used:  the "normal" IPA log buffer; and
the "low level" one.  The first parameter to ipa_ipc_logging() is
the log buffer pointer to use, but we can just as well pass a
Boolean flag to indicate which one we want to write to.  As a result
we can let the logging function take care or tracking the log buffer
pointers, and they don't need to be exposed.  This eliminates the
need for ipa3_get_ipc_logbuf() and ipa3_get_ipc_logbuf_low().

Change-Id: Idf90ad6f9ce3296fe0bea129e8aaffe97303eca8
Signed-off-by: Alex Elder <elder@codeaurora.org>