msm: ipa: change flag meaning in ipa_ipc_logging()
Whenever ipa_ipc_logging() is called, it's either called once to log
something to the "low level" log only, or it's called twice in
succession to log to the "normal" log and then the "low level" log.
Change the flag passed to ipa_ipc_logging() so it inicates whether
both or only the low-level log should be written to, and have it
take care of writing both logs in the same call.

Change-Id: I291b26534d7155b2dc4db09e0de634a628705723
Signed-off-by: Alex Elder <elder@codeaurora.org>