msm: ipa: use IPA logging macros for IPA HAL code
Replace calls to IPAHAL_DBG(), IPAHAL_DBG_LOW(), and IPAHAL_ERR(),
and with calls to ipa_dbg(), ipa_dbg_low(), and ipa_err(),
respectively.  The result is identical, but now the code uses the
common ipa_*() logging macros defined in "ipa_common_i.h".  With the
new macros, the pr_fmt() macro defined at the top of a source file
determines how messages are customized when logged.

This leaves the macro ipa_ipc_logging() no longer used, so it can be
deleted.

Change-Id: I1693a33cad5f45c59764f88ed60af12a068f0fc8
Signed-off-by: Alex Elder <elder@codeaurora.org>