msm: ipa: get rid of IPA_EP_NOT_ALLOCATED
IPA_EP_NOT_ALLOCATED is used for one purpose:  ipa3_get_ep_mapping()
returns it to indicate an error, as opposed to a (non-negative) mapping
value.  Endpoint numbers are never negative, so returning any negative
value from ipa3_get_ep_mapping() is sufficient to indicate an error.

The most typical way for a Linux function to return an error is to
return a negative errno.  Using IPA_EP_NOT_ALLOCATED doesn't fit
that pattern, and the conventional way works fine in this case.

So just have "negative is not valid" be the rule for endpoint
numbers.  Have ipa3_get_ep_mapping() return a negative errno, and
get rid of IPA_EP_NOT_ALLOCATED.

Change-Id: I6361ad7206e44368ec9d6b70add4becee2f1b926
Signed-off-by: Alex Elder <elder@codeaurora.org>