msm: ipa: get rid of ipa3_context->hdr_proc_ctx_tbl
The hdr_proc_ctx_tbl field of the ipa3_context structure is never
used once initialized.  So get rid of it and the code that
initializes it.  That makes the ipa3_hdr_proc_ctx_tbl structure
unused, so it can go away too, as can its type, struct ipa3_hdr_tbl.

Change-Id: I980ed2f04beb756fcfdbc341176d4708e4a864c7
Signed-off-by: Alex Elder <elder@codeaurora.org>