msm: ipa: get rid of route and filter flags
The hdr_tbl_lcl field of the ipa3_context structure is never used
once initialized in the IPA lite code.  Another set of Boolean flag
fields in the IPA context are never initialized either, so can be
assumed to be false.

The only time any of these is used is in logging some messages to
report their values.  Get rid of the fields and related code.
Just dump their known values (0) in the log messages.

Change-Id: I890bb94e5e64a44983ecf0e76e259d3e7327f212
Signed-off-by: Alex Elder <elder@codeaurora.org>