msm: ipa: introduce const_string_read_fop()
Introduce a new file operation function intended to implement the
read of a read-only string value.  This would be better included
in the main debugfs code (in "fs/debugfs/file.c") but we'll wait
until we're getting this stuff upstream to propose that.

This will be used in the next patch; until it's used add a void
reference to the file operations global to avoid a compile error.

Change-Id: I3865f55930812e5f10b48bc7d9f6384c366e4111
Signed-off-by: Alex Elder <elder@codeaurora.org>