msm: ipa: clock scaling is enabled
The enable_clock_scaling field of the ipa3_context structure is
initialized to 1 in ipa3_pre_init() and never changed thereafter.
Change the code to assume it's value is 1, and delete the field in
the IPA context.

Show a constant string "1" in the debugfs file that previously
showed the value of this field.  It's possible this can go away, but
until I'm sure I'll preserve the current functionality.

Change-Id: I78463381b86e04b9299b016f48d26ca99f9a3f04
Signed-off-by: Alex Elder <elder@codeaurora.org>