msm: ipa: remove ipa3_context->hw_stats
The hw_stats field of an IPA context structure is never referenced
in the IPA lite code, so it can be removed.  This removes the only
use of the ipa_hw_stats structure, so its definition can be removed.
And that removes the last use of the ipa_hw_stats_drop structure.

Note that in addition to the one in the ipa_hw_stats_drop, there is
one other reference to the ipa_drop_stats_all structure; but that is
a declaration for a non-existent function.  So delete that function
declaration, along with two related ones that do not exist.

That makes the ipa_hw_stats_drop structure no longer needed, which
makes the ipa_drop_stats structure no longer needed either, so get
rid of them.

Change-Id: I5b560d01012762a18529bce904dcca69a1861c28
Signed-off-by: Alex Elder <elder@codeaurora.org>