msm: ipa: get rid of struct ipa_quota_stats_all
The ipa_quota_stats_all structure is only used in the declaration
for ipa_get_quota_stats(), which is a non-existent function in the
IPA lite code.  Delete that function the declaration, as well as the
declarations of some some related non-existent functions.  Delete
the definition of the ipa_quota_stats_all structure, as well as
the (consequently unused) ipa_quota_stats structure.

Change-Id: Id4d6cac12fc33990d78f12945952f159c1780c3f
Signed-off-by: Alex Elder <elder@codeaurora.org>