msm: ipa: remove ipa_ssr_event enumerated type
The ipa_ssr_event enumerated type is never used in the IPA lite
code.  Its IPA_SSR_EVENT_MAX value (the number of SSR event types)
is used to initialize IPA_EVENT_MAX_NUM as the number of *all*
event types.

The only use of the other ipa_ssr_event values is to create a table
containing human-readable names to represent them.  For each of
these (non-MAX) values, we can safely delete it along with its entry
in the human-readable string table.

With entries in the string table, we can delete it.  The table is
referenced in formatting a string in ipa3_read_msg(), but no event
value will refer to a valid table entry, so just get rid of those
references.

That leaves IPA_EVENT_MAX_NUM equal to 0; so we can just use 0 to
initialize IPA_EVENT_MAX_NUM, and delete the enumerated type
entirely.  (The next patch will clean up this up further.)

Change-Id: I24cba56d4d3704269b1b86c3e9872bd64553c433
Signed-off-by: Alex Elder <elder@codeaurora.org>