msm: ipa: remove the IPA context event arrays
The only use of the msg_w[] and msg_r[] arrays in the IPA context
is in dumping their values via "/sys/kernel/debug/ipa/msg".  They
are never initialized nor are they modified in the IPA lite code, so
they can be assumed to have value 0.

Delete the arrays, and hard-code all the values (0) shown when
accessing the ".../ipa/msg" debugfs file.

Change-Id: I32e01986859beecd1de6e0ffb61120c5845f6a9d
Signed-off-by: Alex Elder <elder@codeaurora.org>