msm: ipa: get rid of /sys/kernel/debug/ipa/msg
IPA_EVENT_MAX_NUM is 0, and as a result, the loop in ipa3_read_msg()
will format nothing into the buffer to be displayed to the user.  We
can therefore delete the loop.  That leaves 0 bytes to be passed to
the user in simple_read_from_buffer(), so that function will return
immediately with value 0.  We can instead have ipa3_read_msg() just
be a trivial function that returns 0.

That's a reasonable change.  But it begs the question of the purpose
served by the "msg" debugfs file...  In other cases I've attempted to
preserve all existing functionality, but in this case I'm going to
just delete the useless file and all its supporting code.

And with that, we can remove the definition of IPA_EVENT_MAX_NUM.

Change-Id: Ieca920ef6f73bd7cfca5415ba5fc1d658e5eb477
Signed-off-by: Alex Elder <elder@codeaurora.org>