msm: ipa: get rid of ipa_smmu_cb_ctx->va_size
The va_size field in the ipa_smmu_cb_ctx stores the size of an
address range.  The end of the range is also stored in the SMMU
context.  Other than computing the value of va_end, the va_size
field is only used in creating the IOMMU mapping; that occurs in
the same function as fetches the va_size value from the DTB.

There's no need to store both va_size and va_end.  The latter is
used in a couple other spots in the code, so we'll choose to keep
that one.  Use a local variable to hold the fetched va_size value
and stop recording it in the SMMU context.

Change-Id: I87ee5322cc04aef3a75d73dfcbb98145cf16ac1f
Signed-off-by: Alex Elder <elder@codeaurora.org>