msm: ipa: move debug print into ipa3_iommu_map()
There are two places that call ipa3_iommu_map(), and each call is
immediately preceded by an identical pr_debug() call.  Move that
call into ipa3_iommu_map() so it's common code.

One consequence of this is that the function name and line number of
the calls will be shared, which means there's know way to know which
caller is responsible for the call.  If this is important
information we should handle this differently.

Change-Id: I043ba83898c654ca390f41d2f124e1bcc49c82d4
Signed-off-by: Alex Elder <elder@codeaurora.org>