msm: ipa: introduce ipa_smmu_probe_common()
A lot of code in the two SMMU probe functions is essentially the
same.  This patch introduces a new function ipa_smmu_probe_common(),
which factors out some code common to both functions.  The only
thing lost is that the debug messages no longer identify which
node is involved.  If this is important we can address it
separately.

Change-Id: I2acb7779dc2df53fca839d92ddbccdaa41eecc4e
Signed-off-by: Alex Elder <elder@codeaurora.org>