msm: ipa: finish expanding ipa_smmu_probe_common()
Put one last block of code into ipa_smmu_probe_common().  The result
makes ipa_smmu_uc_cb_probe() an almost trivial function, using the
common routine to do essentially all of the work.

This change preserves the only behavior that differed in the code
that's now common:  if an error occurs in arm_iommu_attach_device(),
the probe for the AP SMMU CB just reports it with a debug message,
while the other probe reports it as an error, and releases the IOMMU
mapping.  (Whether this difference is correct is something to be
considered  separately.)

Change-Id: I47913bb4c288388f55e5d9ae29f58aebc8c3fd2d
Signed-off-by: Alex Elder <elder@codeaurora.org>