msm: ipa: set DMA masks later in ipa_smmu_attach()
If we are unable to create a mapping in ipa_smmu_attach(), we return
EPROBE_DEFER assuming the failure was due to the IOMMU driver not
being ready.  Avoid repeatedly setting the DMA masks in case of
deferral by setting it later in that function.

Use dma_set_mask_and_coherent() rather than making two separate
calls, since we're setting both masks to the same value.

Change-Id: I7dbbfbbb4249c34e3788e5c8824b60aeebe98005
Signed-off-by: Alex Elder <elder@codeaurora.org>