msm: ipa: allocate ipa3_bus_scale_table after ipa3_ctx
This patch and patches that follow will continue to move the
allocation of ipa3_ctx earlier, with the general reasoning that
it's harmless because--basically by definition--the IPA context is
never used prior to the point it was before the move.  But there's
more to it than that.

Here, we are moving initialization of ipa3_bus_scale_table to be
after the IPA context is allocated, and it is a *global* variable.
Moving its initialization after that of ipa3_ctx is only valid if no
code touches ipa3_bus_scale_table before ipa3_ctx becomes non-null.

A long series of recent patches provided proof that many places in
the code were checking for a null ipa3_ctx pointer unnecessarily.
The only places that remain that check for a null pointer are those
that *must* check; everywhere else can safely assume it's non-null.

So to be valid, we need to ensure places that wait for ipa3_ctx to
be non-null don't use ipa3_bus_scale_table once set; or put another
way, we need to ensure ipa3_ctx is already non-null everywhere
ipa3_bus_scale_table is used.

Other than its initialization in ipa3_bus_scale_table(), the only
place ipa3_bus_scale_table is used is in ipa3_pre_init().  We know
ipa3_pre_init() is only called by ipa_smmu_ap_cb_probe() *after*
ipa3_ctx is allocated.  The first use of ipa3_bus_scale_table
already follows allocation of ipa3_ctx; so there is no harm in
initializing ipa3_ctx earlier than ipa3_bus_scale_table.

Change-Id: I73d9a7f48b3192150696052a19128ea830fe8eaa
Signed-off-by: Alex Elder <elder@codeaurora.org>