msm: ipa_dp: use last_desc for desc_num == 1
In ipa3_send_cmd(), if desc_num == 1, the passed-in descriptor
pointer is typically just the address of a single descriptor.  The
function treats that as if it were a different case than an array of
descriptors.  In fact, it's the same as an array of 1 descriptor.

Move the assignment of last_desc so it applies to the case where
only one descriptor is provided.  And then change the block that
depends on num_desc == 1 to use that pointer rather than the
passed-in descr pointer.

Change-Id: I4f8579a8f2b1d4e93a8d7dfbe74627f62d7b2caf
Signed-off-by: Alex Elder <elder@codeaurora.org>