msm: ipa_dp: further simplify ipa3_send_cmd()
If only one command is in the descriptor array, ipa3_send_one() is
used to send the commands to the IPA; otherwise ipa3_send() is used.

Since ipa3_send_one() is just a trivial wrapper around ipa3_send(),
there's no real benefit to handling the one-command case specially.

So just get rid of the conditional block, and call ipa3_send()
in ipa3_send_cmd() regardless of the number of commands sent.

Change-Id: Icc61bb1e0417f730ec2a0ac4a57f7250b9e4bbaf
Signed-off-by: Alex Elder <elder@codeaurora.org>