msm: ipa: check for result < 0 for ipa3_get_ep_mapping()
Currently ipa3_get_ep_mapping() returns IPA_EP_NOT_ALLOCATED if
there is an error, and the value of that is -1.  The calling code
sometimes checks whether the returned value is IPA_EP_NOT_ALLOCATED
but more often it just checks for -1.

Change all callers of ipa3_get_ep_mapping() so they treat *any*
negative value as an error.  This prepares for the next patch, which
has that function return a negative errno instead.  (It also removes
the inconsistency use of checking for IPA_EP_NOT_ALLOCATED and -1.)

Change-Id: Ie720402cd50194d1cbd6416bc56bb092897bd7d5
Signed-off-by: Alex Elder <elder@codeaurora.org>