msm: ipa_dp: use last_desc for desc_num == 1
In ipa3_send_cmd_timeout(), if desc_num == 1, the passed-in
descriptor pointer is just the address of a single descriptor.
The function treats that as if it were a different case than an
array of descriptors.  In fact, it's the same as an array of 1
descriptor.

Move the assignment of last_desc so it applies to the case where
only one descriptor is provided.  And then change the block that
depends on num_desc == 1 to use that pointer rather than the
passed-in descr pointer.

Change-Id: I1a57689fe50f79520e2122faae5640a6eec6c453
Signed-off-by: Alex Elder <elder@codeaurora.org>