msm: ipa: map modem SMEM memory after ipa3_ctx allocation
SMEM for the modem filter table is allocated and mapped in
ipa_smmu_ap_cb_probe().  This patch moves that initialization
so it occurs after ipa3_ctx has been allocated.  To be a valid
change, other code involving the mapped SMEM memory must execute
only after ipa3_ctx is known to be non-null.

Allocating and mapping this SMEM happens very early.  No activity
involving the modem will occur before ipa3_post_init() runs, and
that is known to only run when ipa3_ctx is valid (and in fact, only
when IPA is in READY state).

So moving the allocation of ipa3_ctx up in this was is harmless.

Change-Id: I40ed00a600cb08be3e0d0d640a99ab39548e27aa
Signed-off-by: Alex Elder <elder@codeaurora.org>