msm: ipa: allocate ipa3_ctx earlier in ipa_smmu_ap_cb_probe()
Move the allocation the IPA context earlier, to the beginning of
ipa_smmu_ap_cb_probe()--before ipa_smmu_attach() is called for the
AP SMMU context block.

The mapping created by ipa_smmu_attach(), and attached to the
device, will not be used until ipa3_post_init(), which we know is
called only after ipa3_ctx is non-null.

Nothing is affected by the mapping attached by ipa_smmu_attach(),
so it's harmless to allocate the IPA context before that call.

Change-Id: Ib4ec04b72acd41bc6aeb7c94d9fb179c1b44c180
Signed-off-by: Alex Elder <elder@codeaurora.org>