msm: ipa: fix bus scale table cleanup
In ipa_smmu_ap_cb_probe() the global ipa3_bus_scale_table is
initialized by a call to msm_bus_cl_get_pdata().  The global
is used to initialize a field in the IPA controller field of
an IPA context.

If an error occurs ipa3_pre_init(), the cleanup path is calling the
cleanup for the *global* bus scale table.  This is unbalanced.

Instead, ipa3_pre_init() should just clear the IPA controller field
it assigned in its cleanup path.  The caller, ipa_smmu_ap_cb_probe(),
is the one that initialized the global variable and should be
responsible for cleaning it up in the event of an error.

Have ipa3_bus_scale_table get cleaned up in ipa_smmu_ap_cb_probe()
if an error occurs in ipa3_pre_init().

Change-Id: I6e3c0dea5c369483769f7d4e814439874779f20e
Signed-off-by: Alex Elder <elder@codeaurora.org>