msm: ipa: allocate IPA context in ipa3_plat_drv_probe()
The IPA context is a global pointer that is used by all of the IPA
code.  It really is "owned" by the top-level driver (whose DTB node
matches "qcom,ipa"), but the AP SMMU context block driver (matching
"qcom,ipa-smmu-ap-cb") is what now allocates it.  This patch moves
the allocation of the IPA context so it occurs when the top-level
node is probed.

The single ipa3_plat_drv_probe() function effectively acts as a set
of distinct probe routines.  The match table for the top-level IPA
platform driver contains five entries.  In addition to matching the
top-level ("qcom,ipa") node, it is used to match children of the
top-level IPA DTB node.  The probe function looks at the compatible
string that matched, and if it's one of the children, a separate
function is called to perform probe of the child node.

When the device tree is parsed at system initialization time, the
top-level IPA node is found, and the probe function gets called.
The compatible string will match none of the child node strings,
so the top-level initialization proceeds.  When this is done,
of_platform_populate() is called, to parse child nodes of the
top-level node, supplying the same match table.

As a result, children of the top-level IPA node are matched, and
for each, a new platform device is created and the common probe
routine is called.

The AP SMMU context block driver ("qcom,ipa-smmu-ap-cb") will get
probed at this time--as a result of ipa3_plat_drv_probe() calling
of_platform_populate().

This patch moves the allocation of ipa3_ctx out of the top of
ipa_smmu_ap_cb_probe(), and into ipa3_plat_drv_probe() just
prior to the call to of_platform_populate().  This means the
allocation occurs at effectively the same time.

As a precaution, we add a check for a null IPA context pointer
at the top of ipa_smmu_ap_cb_probe(); if it's null, -EPROBE_DEFER
is returned, so the probe of this child node will be attempted
again.

Change-Id: Iadb92ca7c00add23dbbbe3506b3467f2c1b4a9ff
Signed-off-by: Alex Elder <elder@codeaurora.org>