msm: ipa: use a preallocated IPA context
Use a statically-defined structure to hold the IPA context.  Don't
update all the references to ipa3_ctx; just have that pointer refer
to the static structure.  This eliminates a possible (though unlikely)
error in the top-level probe routine, and makes it unnecessary to
free or reinitialize the pointer in error paths.

Change-Id: I29402b7fd1204462909cf47ba398ffebec1dafa1
Signed-off-by: Alex Elder <elder@codeaurora.org>