msm: ipa: initialize log buffer early
The IPA log buffer isn't created until ipa3_pre_init().  This was
probably delayed until that time because that's where the IPA
context used to be allocated.

Now that we've got that allocated early, we can move initialization
of other things early.  Start with the log context.  Initializing
that right away means we can log some additional information early.
Move the initialization of the main platform device pointer up too,
so it's the first thing done by the top-level probe function.

Note that we now clean up the log buffer if an error occurs after it
has been created; previously if fetching the "ipa-base" property
failed this did not occur.

Change-Id: I94b4401d4f421c4044891b352b3871d3d180001c
Signed-off-by: Alex Elder <elder@codeaurora.org>