msm: ipa: fetch "qcom,ee" property early
The execution environment (EE) number is a fundamental property of
the AP's interface to the IPA.  It should be initialized when the
top-level driver is initialized, not in ipa3_pre_init().  (Now that
the IPA context is initialized early we can do this.)

Note that we now zero out the "ee" field in case an error occurs
after it has been assigned; this did not previously occur.

Change-Id: Idd17d40706a83028c1a5d3c3f44a1589586435ba
Signed-off-by: Alex Elder <elder@codeaurora.org>