msm: ipa: replace all printk() calls
Replace all printk() calls with ipa_err(), ipa_info(), or ipa_debug()
calls instead.  The log buffer is allocated early now, so we can
assume all messages can be logged.

One call was wrapped by #ifdef DEBUG.  The reason was that pr_debug()
calls are supposed to expand to no code if DEBUG is not defined.
Now that we call ipa_debug(), that macro will take care of not
emitting a printk() in that case.  We'll still send debug messages
to the IPA log though.

In some cases, this is logging in more places than before--some
calls to ipa_debug() were previously just pr_debug() calls rather
than IPADBG().  This is basically a policy decision to always use
the IPA logging calls for messages.

Change-Id: I51857caadd0a14b8d14e1c210c47ba23afb546cd
Signed-off-by: Alex Elder <elder@codeaurora.org>