msm: ipa: skip ipa3_active_clients_logging.log_rdy init
Only ipa_smmu_ap_cb_probe() calls ipa3_pre_init(), and for a proper
DTB that only happens once.  The IPA context is already zero-filled
at startup time, so it's not strictly necessary to initialize any
fields to 0/NULL/false.  The line that initializes the log_rdy field
of the ipa3_active_clients structure embedded in the IPA context to
false is therefore not needed.  So get rid of it.

Change-Id: I7eb626b5a4b4eeb1c770e52b0bcdced23925bd28
Signed-off-by: Alex Elder <elder@codeaurora.org>