msm: ipa: move the definition of struct ipa3_controller
The ipa3_controller structure is below the ipa3_context structure in
"ipa_i.h".  Move it up, so the structure definition (not just a
pointer) can be used to embed the controller structure inside the
IPA context.  Because the ipa3_controller structure refers to it,
this means the ipa3_mem_partition structure definition has to be
moved up as well.

Change-Id: Ice02400eb0c3a59c34c13221a52b9b0d3def0000
Signed-off-by: Alex Elder <elder@codeaurora.org>