msm: ipa: embed the IPA controller struct
An IPA has exactly one controller structure, which is of fixed size.
There's no need to allocate that structure dynamically; just embed
it inside the IPA context instead.

For now, preserve the ipa3_context->ctrl pointer (to avoid touching
lots of code) and initialize it to point to the address of the new
embedded structure rather than allocating it.

Change-Id: I583bf882919a015934373527c1750fcbd014a6af
Signed-off-by: Alex Elder <elder@codeaurora.org>