msm: ipa: ipa3_controller_static_bind() won't fail
All ipa3_controller_static_bind() does is initialize the fields in
the IPA controller structure it's provided.  It won't ever return an
error.  Change its return type to void to reflect that, and get rid
of the error handling code in its caller.  Fix some minor formatting
problems in that function too.

There is no "un-do" function for ipa3_controller_static_bind(); but
we know what it does, so for now zero out the structure in the event
of an error.  Note that this did not previously occur if an error
occurred after it was initialized.

Change-Id: Ibe02b275476024ec6eb5f5f3e4f3e9e158b50113
Signed-off-by: Alex Elder <elder@codeaurora.org>