msm: ipa: initialize memory partition earlier
All ipa3_init_mem_partition() does is fetch an array describing the
IPA SMEM memory partition layout, validate the values found, and
store the result in the IPA controller mem_partition array.  This
can be done by the main IPA driver, during early initialization in
its the probe routine.

Note that there is no un-do function for initializing the memory
partition array.  By the zeroing of the control information that we
already do in the error path we are achieving this effect; this was
not done previously.

Change-Id: Icb03b16fca94d6c89686435021ef1df4f00364b0
Signed-off-by: Alex Elder <elder@codeaurora.org>