msm: ipa: eliminate ipa3_context->bus_scale_table
In ipa_smmu_ap_cb_probe(), ipa3_context->bus_scale_table is
initialized, and if it's no-null, its value is assigned to the
bus_scale_table field in the IPA context controller structure.

There is no other use for the bus_scale_table field in the main IPA
context structure.  So just assign the value directly into the
controller's field and get rid of the copy in the IPA context.

Change-Id: I4215a1223365e6e54cd2a714492e54392a60b93d
Signed-off-by: Alex Elder <elder@codeaurora.org>