msm: ipa: initialize active clients log in main driver
All ipa3_active_clients_log_init() does is initialize the
ipa3_active_clients_logging structure embedded in the IPA
context.  It allocates some zeroed buffers, so a subsequent
error must call ipa3_active_clients_log_destroy() to clean
up the effects.  But otherwise there's no harm in doing this
initialization in the main IPA driver, in the probe routine.

Fix a comment that refers to "getting ipa-clk."  It should have been
updated long ago, with this commit, which removed the ipa3_clk
because it was not used:
  msm: ipa: get rid of ipa3_clk pointer

Change-Id: I7b3cb6f9e1706b381785ba83586db631ddb43d19
Signed-off-by: Alex Elder <elder@codeaurora.org>