msm: ipahal: base and dev are good in ipahal_init()
Only ipa3_pre_init() calls ipahal_init().  That location ensures the
base address passed is non-null immediately before the call, so we
don't need to check that.

The dev pointer passed is also non-null.  It comes from
ipa3_ctx->ap_smmu_cb.dev, which is assigned by ipa_smmu_attach() to
be the dev value passed to ipa_smmu_ap_cb_probe().  That value comes
from ipa3_plat_drv_probe(), and is the address of an object (the
embedded device structure in the platform device for the AP SMMU
context bank platform driver).  So we don't need to check for that
either.

Change-Id: I97dcdffe20a8c742d19d03750df3d1ed63ef4326
Signed-off-by: Alex Elder <elder@codeaurora.org>