msm: ipahal: don't put zero_obj on the stack
In ipahal_reg_init() there's an object defined whose purpose is just
to be zero-filled, for comparison against other objects.  Give that
object static lifetime rather than putting it on the stack.  This
avoids consuming the stack space and has the side-effect of being
initially zero-filled.  Make it const as well.

Change-Id: Ib09421d092b31901408f61d8bf950c7de74d4153
Signed-off-by: Alex Elder <elder@codeaurora.org>