msm: ipahal: populate a single regs structure
If a system has an IPA, it has exactly one of them, and it will have
a single hardware version/type.  There is no need to populate the
entire array of register objects for all hardware versions.  Instead,
we can use a single block of registers, which can be built up using
the ipahal_reg_objs[][] array.  We still verify every entry in this
register block, but only those used are checked.

For now we just populate this array; this commit temporarily
includes a test to verify the result is the same as what the
existing code produces for the current version as well.

Change-Id: I424a34037044508b9f33c28fed70d01e13fc3f0c
Signed-off-by: Alex Elder <elder@codeaurora.org>