msm: ipahal: make ipahal_reg_objs[][] constant
We now use the ipa_regs[] array throughout the HAL code.  The
ipahal_reg_objs[][] array is used only for initializing that, and
this is done prior to populating zero entries found therein.  We no
longer need to update zero entries, so stop doing that.

That means ipahal_reg_objs[][] is no longer ever modified, so we can
declare its content constant.

Change-Id: Ib9857dae02284bc884f34d9da7a2241f6c0066da
Signed-off-by: Alex Elder <elder@codeaurora.org>