msm: ipahal: define and use OFFSET_INVAL
Currently an offset value of -1 in an ipahal_reg_objs[][] entry
represents a register that is not valid for a particular version of
IPA hardware.  This patch defines symbolic constant OFFSET_INVAL for
that purpose.  The value used is the same, but it's defined as an
explicitly unsigned all-1's value rather than having -1 be
implicitly converted to "offset" field's unsigned type.

Change-Id: I86e6fa1ca5e08f79a0117fab679f42b6277d872b
Signed-off-by: Alex Elder <elder@codeaurora.org>