msm: ipahal: 0 offset means uninitialized entry
No valid entry in the ipahal_reg_objs[][] array has a zero offset
value.  Leverage that fact, and use a 0 offset (alone) to indicate
an empty/zero entry.

Update the validation function; it will never be supplied an entry
with a 0 offset.  Also update ipahal_get_reg_mn_ofst() as well as
ipahal_read_reg_*() and ipahal_write_reg_*() to return an error if
ever asked to access a register whose offset is 0.

Change-Id: I1728ef4a0240eb7442f583bd5ca811486010a707
Signed-off-by: Alex Elder <elder@codeaurora.org>