msm: ipahal: kill ipareg_construct_dummy()
There's little value in ipareg_construct_dummy(), which simply
prints a message if it is ever called.  Instead, allow entries in
the ipahal_reg_objs[][] table to provide a null construct function
pointer, and check for null before using it.  If it's null, print
a message and call WARN() just as ipareg_construct_dummy() would
have.

Change-Id: I359321306213c7fa7a4e490073d74c8b620d87da
Signed-off-by: Alex Elder <elder@codeaurora.org>