msm: ipahal: reformat ipahal_reg_objs[][]
Reformat the ipahal_reg_objs[][] array to more obviously show which
groups of definitions constitute sub-arrays.  Each major index value
in the array has a block of initializers for the entries in that
index's sub-array within ipahal_reg_objs[][].

Beyond that, lay out the initializers in a way that's typical in
upstream code.

Change-Id: I7b960ffc9e89f40cfb2f411c7076811c28696380
Signed-off-by: Alex Elder <elder@codeaurora.org>