msm: ipahal: return 0 for errors in ipahal_read_reg_*()
The IPA HAL functions that read registers return an unsigned value.
However in each, there is error checking, and a negative errno is
returned if an error occurs.

None of the callers of these functions actually check for this, and
we can most likely just do away with all of these checks because the
conditions they're checking will never occur.

However, for now, let's at least return a non-negative value.  Have
these functions return 0 on error.  It's no worse, and probably better
than the -EFAULT (-14) as they do right now...

Change-Id: I5aa56908f42ccdfd76611853d89abdb6d0de10a9
Signed-off-by: Alex Elder <elder@codeaurora.org>