msm: ipahal: factor out some code
All of the ipahal_read_reg_*() and ipahal_write_reg_*() functions
compute the offset of the register to read or write.  Remove all
that duplicated code by just using ipahal_get_reg_n_ofst().  This
loses a little bit of debugging information (about which particular
function was computing the offset), but we can address that later if
it's an issue.

Change-Id: I150c72aba230ab9a8283deb79fe309575b68f04d
Signed-off-by: Alex Elder <elder@codeaurora.org>